refactor(media): relocate MediaSupported and lift to ClaimControlledModel
Moves MediaSupported from apps/core/models.py to apps/media/models/base.py
(where the marker class belongs alongside the app that defines what
"media-attachable" means) and changes the base from models.Model to
ClaimControlledModel. media_attachment is itself a claim field, so every
MediaSupported entity is by construction a claim subject — the inheritance
encodes that structural commitment as a compile-time guarantee, which lets
provenance.assert_claim's subject parameter narrow to ClaimControlledModel
without casts at media's three call sites.

The diamond (concrete → CatalogModel → ClaimControlledModel and concrete →
MediaSupported → ClaimControlledModel) is safe today because the only
field on ClaimControlledModel is a GenericRelation, which is routed into
_meta.private_fields and skipped by Django's clash check. A tripwire
docstring on ClaimControlledModel flags that adding any non-private field
would break the diamond and require refactoring first.

AppBoundaries.md amended to permit the targeted media.models →
provenance.models edge (ClaimControlledModel only); the rest of media.models
stays peer-isolated.
