+++ backend/apps/catalog/models/machine_model.py
@@ -10,7 +10,6 @@ from django.db import models
 
 from apps.core.models import (
     MarkdownField,
-    MediaSupported,
     SluggedModel,
     TimeStampedModel,
     field_not_blank,
@@ -19,6 +18,7 @@ from apps.core.models import (
     status_valid,
 )
 from apps.core.validators import validate_no_mojibake
+from apps.media.models import MediaSupported
 
 from .base import CatalogModel
