+++ backend/apps/catalog/models/manufacturer.py
@@ -10,7 +10,6 @@ from django.db.models.functions import Lower
 from apps.core.models import (
     AliasBase,
     MarkdownField,
-    MediaSupported,
     SluggedModel,
     TimeStampedModel,
     field_not_blank,
@@ -19,6 +18,7 @@ from apps.core.models import (
     status_valid,
 )
 from apps.core.validators import validate_no_mojibake
+from apps.media.models import MediaSupported
 
 from .base import CatalogModel
