+++ backend/apps/media/models/__init__.py
@@ -5,7 +5,8 @@ continues to work unchanged.
 """
 
 from .asset import MediaAsset
+from .base import MediaSupported
 from .entity import EntityMedia
 from .rendition import MediaRendition
 
-__all__ = ["EntityMedia", "MediaAsset", "MediaRendition"]
+__all__ = ["EntityMedia", "MediaAsset", "MediaRendition", "MediaSupported"]
